cover-letter: Converge on using secs_to_jiffies() part two
This is the second series that converts users of msecs_to_jiffies() that
either use the multiply pattern of either of:
- msecs_to_jiffies(N*1000) or
- msecs_to_jiffies(N*MSEC_PER_SEC)

where N is a constant or an expression, to avoid the multiplication.

The conversion is made with Coccinelle with the secs_to_jiffies() script
in scripts/coccinelle/misc. Learning from part 1, attention is paid to
what the best change can be rather than restricting to what the tool
provides.

Signed-off-by: Easwar Hariharan <eahariha@linux.microsoft.com>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20241217-converge-secs-to-jiffies-part-two-f44017aa6f67",
    "prefixes": []
  }
}